Move Comment UI functions into separate file commentsUI.js
