+++ src/index.js
@@ -2,6 +2,7 @@ import './style.css';
 import { getLikes, postLikes } from './modules/likes.js';
 import { getComments, postComments, countComments } from './modules/comments.js';
 import count from './modules/itemsCounter.js';
+import { showMessage, getImageAndDescription, getCommentForm } from './modules/commentsUI';
 
 const cars = document.querySelector('.cars');
 const nature = document.querySelector('.nature');
@@ -22,49 +23,19 @@ const asyncGetCall = async (photographer) => {
   }
 };
 
-const showMessage = (msg, status) => {
-  const displayMessage = document.querySelector('.message');
-  displayMessage.classList.add(status);
-  displayMessage.innerHTML = msg;
-
-  setTimeout(() => {
-    displayMessage.innerHTML = '';
-    displayMessage.classList.remove(status);
-  }, 2000);
-};
-
-const getImageAndDescription = (imgObj) => {
-  return `<img src=${imgObj[0].img} class="popup-image">
-      <button class="btn-close" id="close-comment"><i class="fa-solid fa-xmark"></i></button>
-      <h3>${imgObj[0].desc}</h3>
-      <div class="description">
-        <p>width: ${imgObj[0].width}</p>
-        <p>height: ${imgObj[0].height}</p>
-        <p>id: ${imgObj[0].id}</p>
-        <p>color: ${imgObj[0].color}</p>
-      </div>`;
-};
-
-const getCommentForm = () => {
-  const form = document.createElement('form');
-  form.classList.add('comment-form');
-  form.innerHTML = `<h3>Add a comment</h3>
-      <p class="message"></p>
-      <input type="text" class="user-name" required placeholder="Your name">
-      <textarea class="user-comment" cols="30" rows="10" placeholder="Your Insights"></textarea>
-      <button class="btn btn-add-comment">Comment</button>`;
-  return form;
-};
-
 const updateComments = (imgObj, modal, commentForm) => {
   const commentsDiv = document.createElement('div');
   commentsDiv.classList.add('comments-list');
 
   getComments(imgObj[0].id).then((values) => {
-    const numOfComments = countComments(values);
+    const  filter = values.filter(function (ob) {
+      return ob.username != 'Cem oral' && ob.username != 'Berkay akbaş';
+    });
+    const numOfComments = countComments(filter);
+    
     commentsDiv.innerHTML += `<h3>Comments (${numOfComments})</h3>`;
     if (numOfComments > 0) {
-      values.forEach((value) => {
+      filter.forEach((value) => {
         commentsDiv.innerHTML += `<p>${value.creation_date} ${value.username}: ${value.comment}</p>`;
       });
     }
